Allow for singular values and lists of values

+++ src/main/java/io/prometheus/api/Value.java
@@ -7,5 +7,5 @@ import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 @JsonDeserialize(using = ValueDeserializer.class)
 public record Value (
     Date timestamp,
-    int value
+    Double value
 ) {}
